2e commit division du mainMenu
