+++ src/piledger/func.py
@@ -1,41 +1,21 @@
-from piledger.account import Account
-from piledger.transaction import Transaction
-
-
+from .account import Account
+from .transaction import Transaction
+import csv
 
 
 def read_data_file():  # Lire et analyser le fichier data.csv
     data = []
     with open('data.csv', 'r', encoding='utf-8') as file:
-        lines = file.readlines()
-    for i, line in enumerate(lines):
-        if i == 0:
-            continue  # Skip header
-        line = line.strip()
-        if line:
-            parts = []
-            current_part = ""
-            in_quotes = False
-            j = 0
-            while j < len(line):
-                char = line[j]
-                if char == '"':
-                    in_quotes = not in_quotes
-                elif char == ',' and not in_quotes:
-                    parts.append(current_part)
-                    current_part = ""
-                    j += 1
-                    continue
-                current_part += char
-                j += 1
-            parts.append(current_part)
-            if len(parts) >= 5:
+        csv_reader = csv.reader(file)
+        next(csv_reader)  # Skip header
+        for row in csv_reader:
+            if len(row) >= 5:
                 txn = Transaction(
-                    no_txn=int(parts[0]),
-                    date=parts[1],
-                    account=Account(parts[2]),
-                    amount=float(parts[3]),
-                    comment=parts[4]
+                    no_txn=int(row[0]),
+                    date=row[1],
+                    account=Account(row[2]),
+                    amount=float(row[3]),
+                    comment=row[4]
                 )
                 data.append(txn)
     return data
@@ -54,8 +34,6 @@ def get_all_accounts(data):  # Extraire tous les noms de comptes uniques
         if not any(acc.name == account.name for acc in accounts):
             accounts.append(account)
     return [acc.name for acc in accounts]
-
-
 
 def get_transactions_by_date_range(data, start_date, end_date): # Filtrer les transactions par plage de dates
     filtered_transactions = []
@@ -92,12 +70,18 @@ def find_total_expenses(data): # Dépenses totales (hors compte courant et reven
     return total
 
 def export_account_postings(data, account_name, filename): # Exporter les écritures d'un compte vers un fichier CSV
-    with open(filename, 'w', encoding='utf-8') as file:
-        file.write("No txn,Date,Compte,Montant,Commentaire\n")
+    with open(filename, 'w', encoding='utf-8', newline='') as file:
+        csv_writer = csv.writer(file)
+        csv_writer.writerow(["No txn", "Date", "Compte", "Montant", "Commentaire"])
         for transaction in data:
             if transaction.account.name == account_name:
-                line = f"{transaction.no_txn},{transaction.date},{transaction.account.name},{transaction.amount},{transaction.comment}\n"
-                file.write(line)
+                csv_writer.writerow([
+                    transaction.no_txn,
+                    transaction.date,
+                    transaction.account.name,
+                    transaction.amount,
+                    transaction.comment
+                ])
     print(f"Écritures exportées vers {filename}")
 
 def validate_account_name(accounts, account_name): # Validation du nom de compte (insensible à la casse)
@@ -106,93 +90,3 @@ def validate_account_name(accounts, account_name): # Validation du nom de compte
             return account
     return None
 
-def handle_balance_inquiry(data, accounts):  # Choix 1
-    print("\n--- Consultation de solde ---")
-    print("Comptes disponibles:")
-    for account in accounts:
-        print(f"  - {account}")
-    
-    account_input = input("\nEntrez le nom du compte: ").strip()
-    
-    if not account_input:
-        print("Nom de compte invalide!")
-        return
-    
-    validated_account = validate_account_name(accounts, account_input)
-    
-    if validated_account:
-        balance = calculate_balance(data, validated_account)
-        print(f"\nSolde du compte '{validated_account}': {balance:.2f}$")
-        
-    else:
-        print(f"Compte '{account_input}' introuvable!")
-        print("Vérifiez l'orthographe ou choisissez un compte dans la liste.")
-
-def handle_statistics(data): # Choix 5
-    print("\n=== STATISTIQUES FINANCIÈRES ===")
-    
-    total_income = find_total_income(data)
-    total_expenses = find_total_expenses(data)
-    net_worth = total_income - total_expenses
-    
-    print(f"Revenus totaux: {total_income:.2f}$")
-    print(f"Dépenses totales: {total_expenses:.2f}$")
-    print(f"Situation nette: {net_worth:.2f}$")
-    
-    if net_worth > 0:
-        print("📈 Situation financière positive")
-    elif net_worth < 0:
-        print("📉 Situation financière négative")
-    else:
-        print("⚖️  Situation financière équilibrée")
-    
-    largest_expense = find_largest_expense(data)
-    if largest_expense:
-        print(f"\nPlus grosse dépense: {largest_expense.amount:.2f}$ ({largest_expense.account.name})")
-        if largest_expense.comment:
-            print(f"Commentaire: {largest_expense.comment}")
-    
-    current_account_balance = calculate_balance(data, 'Compte courant')
-    print(f"\nSolde du compte courant: {current_account_balance:.2f}$")
-
-def handle_date_search(data): # Choix 7
-    print("\n--- Recherche par période ---")
-    start_date = input("Date de début (YYYY-MM-DD): ").strip()
-    end_date = input("Date de fin (YYYY-MM-DD): ").strip()
-    
-    if not start_date or not end_date:
-        print("Dates invalides!")
-        return
-    
-    filtered_data = get_transactions_by_date_range(data, start_date, end_date)
-    
-    if len(filtered_data) == 0:
-        print(f"Aucune transaction trouvée entre {start_date} et {end_date}")
-    else:
-        print(f"\n{len(filtered_data)} écritures(s) trouvée(s) entre {start_date} et {end_date}:")
-        for transaction in filtered_data:
-            print(f"  {transaction.date} - {transaction.account.name}: {transaction.amount:.2f}$")
-
-def handle_export(data, accounts): # Choix 6
-    print("\n--- Exportation ---")
-    print("Comptes disponibles:")
-    for account in accounts:
-        print(f"  - {account}")
-    
-    account_input = input("\nEntrez le nom du compte à exporter: ").strip()
-    
-    if not account_input:
-        print("Nom de compte invalide!")
-        return
-    
-    validated_account = validate_account_name(accounts, account_input)
-    
-    if validated_account:
-        filename = input("Nom du fichier de sortie (ex: export.csv): ").strip()
-        if not filename:
-            filename = f"export_{validated_account.replace(' ', '_').lower()}.csv"
-        
-        export_account_postings(data, validated_account, filename)
-    else:
-        print(f"Compte '{account_input}' introuvable!")
-
